requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
python-dotenv==1.0.0
dj-database-url==2.1.0
python-dateutil==2.8.2
drf-spectacular==0.27.0